+++ packages/website/CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 ### Removed
 ### Breaking Changes
 
-## [7.0.0] - 2026-07-04
+## [7.0.0] - 2026-07-05
 
 ### Added
 
@@ -23,6 +23,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 ### Changed
 
 - feat: Restructure the sidebar, flattening the single "Setup" section into four top-level sections: "Classification", "Selectors", "Policies" and "Settings".
+- feat: Change links in navigation bar to point to the new "Classifications", "Selectors", and "Policies" sections, instead of the old "Setup" section.
+- feat: Change footer link from "Setup" to "Settings" section.
 - chore: Upgrade Docusaurus from `3.9.2` to `3.10.1`.
 
 ## [6.0.0] - 2026-03-15